google-workspace-new/functions/get_leavers/requirements.txt
functions-framework==3.*
google_api_python_client==2.114.0
protobuf==4.25.3